# docs/api/service-openapi.yaml
openapi: 3.1.0
info:
  title: TrioBase Governed OpenAPI Integration Service
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: /api/v1/openapi
...
security:
  - applicationCredential: []
paths:
...
  /management/routes/{routeId}/versions:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: createRouteVersion
      summary: Create a draft route version
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200': { description: Route version }
  /management/routes/versions/{versionId}:
    get:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: getRouteVersion
      summary: Get a route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Route version }
    put:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: updateRouteVersion
      summary: Update a draft route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Route version }
  /management/routes/versions/{versionId}/publish:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: publishRouteVersion
      summary: Publish a route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Route version }
  /management/routes/versions/{versionId}/deprecate:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: deprecateRouteVersion
      summary: Deprecate a route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Route version }
  /management/routes/versions/{versionId}/archive:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: archiveRouteVersion
      summary: Archive a route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Route version }
  /management/routes/{routeId}/archive:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: archiveRoute
      summary: Archive a route definition
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200': { description: Route archived }
  /management/routes/{routeKey}/resolve-preview:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: previewRouteResolution
      summary: Preview route resolution before release activation
      parameters:
        - $ref: '#/components/parameters/RouteKey'
      responses:
        '200': { description: Resolved route preview }
  /management/routes/versions/{versionId}/releases:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: createRouteRelease
      summary: Create an immutable release snapshot from a published route version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { description: Release snapshot }
  /management/releases/{releaseId}/activate:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: activateRouteRelease
      summary: Atomically activate a release snapshot
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200': { description: Active release }
  /management/routes/{routeId}/rollback:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: rollbackRouteRelease
      summary: Roll back a route to a previous release snapshot
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200': { description: Active release }
  /management/releases/{releaseId}/deprecate:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: deprecateRouteRelease
      summary: Deprecate a release snapshot
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200': { description: Release snapshot }
  /management/releases/{releaseId}/archive:
    post:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: archiveRouteRelease
      summary: Archive a release snapshot
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200': { description: Release snapshot }
  /management/routes/{routeId}/releases:
    get:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: listRouteReleases
      summary: List release snapshots for a route
      parameters:
        - $ref: '#/components/parameters/RouteId'
      responses:
        '200': { description: Release snapshots }
  /management/routes/{routeKey}/active-release:
    get:
      tags: [Routes]
      x-triobase-service: service-openapi
      operationId: getActiveRouteRelease
      summary: Resolve the active release for a route key and environment
      parameters:
        - $ref: '#/components/parameters/RouteKey'
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200': { description: Active release }
...
components:
  securitySchemes:
    applicationCredential:
      type: apiKey
      in: header
      name: X-Client-Credential
      description: Gateway-authenticated application credential plus active subscription and policy.
    callbackSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC or RSA signature validated with timestamp and nonce replay protection.
    gatewayContext:
      type: apiKey
      in: header
      name: X-OpenAPI-Gateway-Secret
      description: Shared gateway/runtime trust secret. Production callers must also send service identity headers.
  parameters:
    RouteKey:
      name: routeKey
      in: path
      required: true
      schema: { type: string }
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema: { type: string }
...
    RouteId:
      name: routeId
      in: path
      required: true
      schema: { type: string }
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema: { type: string }
...
    EnvironmentQuery:
      name: environment
      in: query
      required: true
      schema:
        type: string
        enum: [DEV, TEST, PROD]
...